+++ db/connect.js
@@ -1,15 +1,28 @@
 const sqlite3 = require('sqlite3').verbose();
-const db = new sqlite3.Database('./db/app.db');
+const path = require('path');
+const fs = require('fs');
+
+// 경로 설정
+function pickDbPath() {
+    const candidates = [
+        process.env.DB_PATH,
+        path.join(process.resourcesPath || __dirname, 'db', 'app.db'),
+        path.join(process.resourcesPath || __dirname, 'app.db'),
+        path.join(__dirname, 'app.db'),
+    ].filter(Boolean);
+
+    for (const p of candidates) {
+        try { if (p && fs.existsSync(p)) return p; } catch { }
+    }
+    // 개발용 기본 경로
+    return path.join(__dirname, 'app.db');
+}
+
+const dbFile = pickDbPath();
+
+const db = new sqlite3.Database(
+    dbFile,
+    sqlite3.OPEN_READWRITE | sqlite3.OPEN_CREATE,
+);
+
 module.exports = db;
-
-// // 33개 부서 목록 전체 출력
-// db.all("SELECT * FROM departments", (err, rows) => {
-//   if (err) return console.error(err.message);
-//   console.log(rows); 
-// });
-
-// // 54개 notice 전체 출력
-// db.all("SELECT * FROM notice", (err, rows) => {
-//   if (err) return console.error(err.message);
-//   console.log(rows);
-// });
