--- mail_detail.html
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-
-<head>
-  <meta charset="UTF-8" />
-  <title>메일 상세</title>
-  <link rel="stylesheet" href="style.css" />
-</head>
-
-<body>
-  <!-- GNB -->
-  <header class="gnb">
-    <img src="logo/gnb_Logo.png" alt="로고" class="gnb-logo" />
-    <input type="text" class="search-bar" placeholder="확인하고 싶은 규정 정보를 검색해보세요" />
-    <button class="search-btn"></button>
-    <div class="gnb-right">
-      <div class="alarm-container">
-        <img src="icon/ic_alarm_new_48.png" alt="알림" class="alarm-icon" />
-      </div>
-      <div class="dept-dropdown">
-        <button class="dept-name-toggle">
-          <span class="dept-name">ID</span>
-          <span class="arrow">▼</span>
-        </button>
-        <div class="dept-menu">
-          <button class="logout-btn">로그아웃</button>
-        </div>
-      </div>
-    </div>
-  </header>
-
-  <main class="container">
-    <!-- 좌측 메뉴 -->
-    <aside class="sidebar">
-      <a href="department.html" class="tab">DEPARTMENT</a>
-      <a href="all_mail.html" class="tab">ALL MAIL</a>
-      <a href="important.html" class="tab">IMPORTANT</a>
-    </aside>
-
-    <div class="sidebar-separator"></div>
-
-    <section class="mail-detail" id="mail-detail">
-      <!-- JS에서 내용 동적 삽입 -->
-    </section>
-
-    <div class="sidebar-separator"></div>
-  </main>
-
-  <script src="script.js"></script>
-  <script src="gnb.js"></script>
-  <script src="mail_detail.js"></script>
-</body>
-
-</html>
