+++ package.json
@@ -1,7 +1,31 @@
 {
   "dependencies": {
     "cors": "^2.8.5",
+    "electron-squirrel-startup": "^1.0.1",
     "express": "^5.1.0",
     "sqlite3": "^5.1.7"
-  }
+  },
+  "name": "kkaebee",
+  "version": "1.0.0",
+  "description": "Desktop application that delivers real-time financial regulation alerts via the internal mail system.",
+  "main": "main.js",
+  "devDependencies": {
+    "@electron-forge/cli": "^7.8.3",
+    "@electron-forge/maker-deb": "^7.8.3",
+    "@electron-forge/maker-rpm": "^7.8.3",
+    "@electron-forge/maker-squirrel": "^7.8.3",
+    "@electron-forge/maker-zip": "^7.8.3",
+    "@electron-forge/plugin-auto-unpack-natives": "^7.8.3",
+    "@electron-forge/plugin-fuses": "^7.8.3",
+    "@electron/fuses": "^1.8.0",
+    "electron": "37.2.6"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "electron-forge start",
+    "package": "electron-forge package",
+    "make": "electron-forge make"
+  },
+  "author": "kkaebee",
+  "license": "ISC"
 }
