+++ public/all_mail.html
@@ -1,0 +1,74 @@
+<!DOCTYPE html>
+<html lang="ko">
+
+<head>
+  <meta charset="UTF-8">
+  <title>모든 메일함</title>
+  <link rel="stylesheet" href="style.css">
+</head>
+
+<body>
+  <!-- GNB -->
+  <header class="gnb">
+    <img src="logo/gnb_Logo.png" alt="로고" class="gnb-logo">
+    <input type="text" class="search-bar" placeholder="확인하고 싶은 규정 정보를 검색해보세요">
+    <button class="search-btn"></button>
+    <div class="gnb-right">
+      <div class="alarm-container">
+        <img src="icon/ic_alarm_new_48.png" alt="알림" class="alarm-icon">
+      </div>
+      <div class="dept-dropdown">
+        <button class="dept-name-toggle">
+          <span class="dept-name">ID</span> <!-- 부서명 자동 변경됨-->
+          <span class="arrow">▼</span>
+        </button>
+        <div class="dept-menu">
+          <button class="logout-btn">로그아웃</button>
+        </div>
+      </div>
+    </div>
+  </header>
+
+  <main class="container">
+    <!-- 좌측 메뉴 -->
+    <aside class="sidebar">
+      <a href="department.html" class="tab">DEPARTMENT</a>
+      <a href="all_mail.html" class="tab orange">ALL MAIL</a>
+      <a href="important.html" class="tab">IMPORTANT</a>
+
+      <a href="alarm.html" class="tab selected">ALARM</a>
+    </aside>
+
+    <div class="sidebar-separator"></div>
+
+    <!-- 메일 목록 -->
+    <section class="mail-list">
+      <h1>모든 메일함</h1>
+      <p class="mail-count">전체 00건</p>
+
+      <ul class="mail-items">
+
+        <!-- 이하 동일한 형식으로 추가 -->
+      </ul>
+
+      <div class="pagination">
+      </div>
+
+    </section>
+
+    <div class="sidebar-separator"></div>
+  </main>
+
+</body>
+<script src="script.js"></script>
+<script src="gnb.js"></script>
+<script src="all_mail.js"></script>
+<script>
+  function goToDetail(id) {
+    // location.href = `mail_detail.html#id=${id}`;
+    localStorage.setItem("selectedMailId", String(id));
+    location.href = "mail_detail.html";
+  }
+</script>
+
+</html>
