+++ public/index.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <title>KKaeBee 로그인</title>
     <link href="https://fonts.googleapis.com/css2?family=Inter&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="../login/login.css" />
+    <link rel="stylesheet" href="./index.css" />
 </head>
 
 <body>
@@ -26,5 +26,6 @@
 
             <button type="submit" class="login-btn">로그인</button>
         </form>
-        <script src="../login/login.js"></script>
+        <script src="./index.js"></script>
+    </div>
 </body>
