+++ public/index.js
@@ -1,5 +1,7 @@
 const loginForm = document.getElementById("login-form");
+const usernameInput = document.getElementById("username");
 const passwordInput = document.getElementById("password");
+const loginBtn = document.querySelector(".login-btn");
 
 passwordInput.addEventListener("input", function () {
     //this.value = this.value.replace(/[^A-Za-z0-9]/g, "");
@@ -8,7 +10,7 @@ passwordInput.addEventListener("input", function () {
 loginForm.addEventListener("submit", async function (e) {
     e.preventDefault();
 
-    const name = document.getElementById("username").value.trim();
+    const name = usernameInput.value.trim();
     const password = passwordInput.value.trim();
 
     try {
@@ -19,8 +21,12 @@ loginForm.addEventListener("submit", async function (e) {
         });
 
         if (!res.ok) {
-            const errorData = await res.json();
-            throw new Error(errorData.message || "로그인 실패");
+        let msg = "로그인 실패";
+        try {
+            const err = await res.json();
+            if (err?.message) msg = err.message;
+        } catch {}
+        throw new Error(msg);
         }
 
         const data = await res.json();
@@ -33,9 +39,17 @@ loginForm.addEventListener("submit", async function (e) {
         localStorage.setItem("dept_id", data.department_id);
 
         
-        window.location.href = "../department.html";
+        window.location.href = "./department.html";
     } catch (err) {
-        console.error("로그인 오류:", err.message);
-        alert("로그인에 실패했습니다. " + err.message);
+        console.error("로그인 오류:", err);
+        passwordInput.value = "";
+        alert("로그인에 실패했습니다.");
+        setTimeout(() => {
+        usernameInput.focus();
+        usernameInput.select?.();
+        }, 0);
+    } finally {
+        // 다시 입력 가능하게
+        loginBtn.disabled = false;
     }
-});
+    });
